backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.5.0
openai==1.40.0
langgraph==0.2.14
langchain-openai==0.1.22
langchain-core==0.2.38
azure-ai-textanalytics==5.3.0
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.30.2
httpx==0.27.0
pydantic==2.8.0
datasketch==1.6.5
pandas==2.2.2
python-multipart==0.0.9